don't re-enable depth test at the end of drawloop
to avoid weird thing at ge exit. We could also call DRW_state_reset() to
ensure the openGL states are ok maybe

The profile is not drawn correctly each time but as it is an
experimental branch I merge that into master waiting for better if we
want to work on it. Feel free to change what I did because it's weird
code