Update armatures with the depsgraph
Notes:

- In BL_BlenderDataConversion, the code related to deformers can cause a
crash. I don't comment it for now because i'm not sure it is not used
for other things than armatures update

- In KX_Scene, update_anim_thread_func is not used anymore because:
1) threading was causing conflicts with depsgraph update
2) all the update code (for armatures/non armatures) is now in
UpdateActionManager (GetDeformer()->Update() is not used)

- The profiler for animations is wrong because the armatures update is
done when the depsgraph is "flushed" and the depsgraph is flushed at the
beginning of render loop. Then most of the armatures update time is
included in Raterizer time. Then we can't really profile animations with
bge code for now. Maybe we can do that with something in Blender code